+++ ci4/app/Interfaces/GatewayAddress.php
@@ -1,0 +1,8 @@
+<?php namespace App\Interfaces;
+/**
+ * @property string $type
+ * @property string $value
+ */
+interface GatewayAddress {
+
+}