+++ ci4/app/Interfaces/GatewayAddressList.php
@@ -1,0 +1,7 @@
+<?php namespace App\Interfaces;
+/**
+ * @property GatewayAddress[] $values
+ */
+interface GatewayAddressList {
+
+}