+++ vue/src/core/services/Deploy/models/GatewayAddress.ts
@@ -1,0 +1,14 @@
+/**
+ * Created by ModelParser
+ * Date: 14-07-2026.
+ * Time: 08:53.
+ */
+import {GatewayAddressDefinition} from './definitions/GatewayAddressDefinition';
+
+export class GatewayAddress extends GatewayAddressDefinition {
+
+    constructor(json?: any) {
+        super(json);
+    }
+
+}