+++ vue/src/core/services/Deploy/models/definitions/GatewayAddressDefinition.ts
@@ -1,18 +1,14 @@
 /**
  * Created by ModelParser
  */
-import {Domain} from '../Domain';
-import {GatewayAddress} from '../GatewayAddress';
 import {User} from '../User';
 import {Deletion} from '../Deletion';
 import {BaseModel} from '../BaseModel';
 
-export class GatewayDefinition extends BaseModel {
-    name?: string;
-    gateway_class_name?: string;
-    namespace?: string;
-    domains?: Domain[];
-    gateway_addresses?: GatewayAddress[];
+export class GatewayAddressDefinition extends BaseModel {
+    gateway_id?: number;
+    type?: string;
+    value?: string;
     id?: number;
     created?: string;
     updated?: string;
@@ -30,11 +26,9 @@ export class GatewayDefinition extends BaseModel {
 
     public populate(data?: any, patch = false) {
         if (!patch) {
-            delete this.name;
-            delete this.gateway_class_name;
-            delete this.namespace;
-            delete this.domains;
-            delete this.gateway_addresses;
+            delete this.gateway_id;
+            delete this.type;
+            delete this.value;
             delete this.id;
             delete this.created;
             delete this.updated;
@@ -47,20 +41,14 @@ export class GatewayDefinition extends BaseModel {
         }
 
         if (!data) return;
-        if (data.name != null) {
-            this.name = data.name;
+        if (data.gateway_id != null) {
+            this.gateway_id = data.gateway_id;
         }
-        if (data.gateway_class_name != null) {
-            this.gateway_class_name = data.gateway_class_name;
+        if (data.type != null) {
+            this.type = data.type;
         }
-        if (data.namespace != null) {
-            this.namespace = data.namespace;
-        }
-        if (data.domains != null) {
-            this.domains = data.domains.map((i: any) => new Domain(i));
-        }
-        if (data.gateway_addresses != null) {
-            this.gateway_addresses = data.gateway_addresses.map((i: any) => new GatewayAddress(i));
+        if (data.value != null) {
+            this.value = data.value;
         }
         if (data.id != null) {
             this.id = data.id;