+++ vue/src/core/services/Deploy/models/index.ts
@@ -32,6 +32,7 @@ export {Domain as Domain} from "./Domain";
 export {EmailService as EmailService} from "./EmailService";
 export {EnvironmentVariable as EnvironmentVariable} from "./EnvironmentVariable";
 export {Gateway as Gateway} from "./Gateway";
+export {GatewayAddress as GatewayAddress} from "./GatewayAddress";
 export {InitContainer as InitContainer} from "./InitContainer";
 export {InitContainerEnvironmentVariable as InitContainerEnvironmentVariable} from "./InitContainerEnvironmentVariable";
 export {K8sCronJob as K8sCronJob} from "./K8sCronJob";